+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * **Add**: Add `shared: true` option to several testing-related calls to enable shared storage of mocks and calls across process boundaries.
   See the **Async and Sharing** section of the `GitHub.Testing` module documentation.
+* **Fix**: Remove empty schemas that represented open map types.
 
 ### 0.3.1 (2024-03-26)