+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+_Nothing yet._
+
+### 0.3.2 (2024-04-25)
+
 * **Add**: Add `shared: true` option to several testing-related calls to enable shared storage of mocks and calls across process boundaries.
   See the **Async and Sharing** section of the `GitHub.Testing` module documentation.
 * **Fix**: Remove empty schemas that represented open map types.